% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demos}
\alias{demos}
\alias{demoNumericRange}
\title{And example showing how numericRangeInput works}
\usage{
demoNumericRange(example = "datepicker")
}
\description{
And example showing how numericRangeInput works
}
\examples{
\dontrun{

if (interactive()) {

demoNumericRange()

}

}
}
