% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRange.R
\docType{package}
\name{shinyRange}
\alias{shinyRange}
\alias{shinyRange-package}
\title{shinyRange: Text-box-based numeric range input widget for Shiny.}
\description{
Text-box-based numeric range input widget for Shiny.
}
